adding the dry-run feature

+++ organizer.py
@@ -18,7 +18,7 @@ file_types = {".jpg" : "Images" ,
               ".html" : "Code" ,
               }
 
-def organize_folder(folder_path_str) :
+def organize_folder(folder_path_str,dry_run) :
     folder_path = Path(folder_path_str)
     if folder_path.is_dir():
         for item in folder_path.iterdir() :     
@@ -31,12 +31,19 @@ def organize_folder(folder_path_str) :
                     if not folder_name : 
                         folder_name= "Other" 
                 target_folder = folder_path / folder_name
-                target_folder.mkdir(exist_ok=True)
-                print("Moving",item.name ,"→",folder_name)
-                shutil.move(item,target_folder)
-            
+                if not dry_run :
+                    target_folder.mkdir(exist_ok=True)
+                    print("Moving",item.name ,"→",folder_name)
+                    shutil.move(item,target_folder)
+                else:
+                        print("Would move",item.name ,"→",folder_name)
+
 if __name__ == "__main__":
     if len(sys.argv) < 2:
         print("Usage: python organize.py <folder_path>")
     else:
-        organize_folder(sys.argv[1])
+        if "--dry-run" in sys.argv : 
+            dry_run = True
+        else:
+            dry_run = False 
+        organize_folder(sys.argv[1],dry_run)
